==> Turkish/02. Hard-rule evaluation.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm

from utils.hard_rules import hard_rules

### Load data

# # Import data
# df = pd.read_csv('Data/_20230622-132905_training.tsv', sep='\t', header=None)

# # Rename columns
# df.rename(columns = {0: 'Value',
#                      1: 'Evidence_type',
#                      2: 'CCV'}, 
#           inplace = True)

# # Remove ENG instances
# ENG_ccvs = [ccv for ccv in df['CCV'].unique() if 'ENG' in ccv]
# df = df[ ~df['CCV'].isin(ENG_ccvs) ]

# # Remove duplicates
# df.drop_duplicates(inplace=True)

# df = df.reset_index().drop('index', axis=1)

# Import data
df = pd.read_csv('Data/_20230622-132905_training.tsv', sep='\t', header=None)

# Rename columns
df.rename(columns = {0: 'Value',
                     1: 'Evidence_type',
                     2: 'CCV'}, 
          inplace = True)


# Remove duplicates
df.drop_duplicates(inplace=True)

# Selected CCVs, which can be handled with hard-rules
selected_CCVs = ['CCV:00004', 'CCV:00012', 'CCV:00013', 'CCV:00014', 
                 'CCV:00071', 'CCV:00072', 'CCV:00073', 'CCV:00084', 
                 'CCV:00094', 'CCV:00034', 'CCV:00035', 'CCV:00036', 'CCV:00100',
                 'CCV:00042', 'CCV:00043', 'CCV:00047', 'CCV:00048', 
                 'CCV:00066', 'CCV:00067', 'CCV:00068', 'CCV:00069', 
                 'CCV:00083', 'CCV:00028', 'CCV:00029', 'CCV:00079', 
                 'CCV:00080', 'CCV:00081', 'CCV:00082', 'CCV:00085', 
                 'CCV:00087', 'CCV:00088', 'CCV:00089', 'CCV:00090', 
                 'CCV:00092', 'CCV:00093', 'CCV:00038', 'CCV:00022', 
                 'CCV:00025', 'CCV:00026', 'CCV:00041', 'CCV:00053', 
                 'CCV:00054', 'CCV:00049', 'CCV:00057', 'CCV:00058', 
                 'CCV:00059', 'CCV:00061', 'CCV:00062', 'CCV:00063',
                 'CCV:00096', 'CCV:00097', 'CCV:00098', 'CCV:00099']

df = df[ df['CCV'].isin(selected_CCVs) ]

df = df.reset_index().drop('index', axis=1)


### Preprocess

# Cannot be classified: 'CCV:00004', 'CCV:00025', 'CCV:00028', 'CCV:00029', 'CCV:00035', 'CCV:00036', 'CCV:00042', 'CCV:00043'
# 'CCV:00047', 'CCV:00048', 'CCV:00049', CCV:00072, 'CCV:00084', 'CCV:00087', 'CCV:00088',
# 'CCV:00079', 'CCV:00082', 'CCV:00092', 'CCV:00093', 'CCV:00096'

from tqdm import tqdm

for i in tqdm( range(df.shape[0]) ):
    text = df['Value'].iloc[i]
    Evidence_type = df['Evidence_type'].iloc[i]
    CCV = df['CCV'].iloc[i]

    if CCV in ['CCV:00004', 'CCV:00025', 'CCV:00035', 'CCV:00036', 'CCV:00042', 'CCV:00043',
    'CCV:00047', 'CCV:00048', 'CCV:00049', 'CCV:00072', 'CCV:00084', 'CCV:00087', 'CCV:00088',
    'CCV:00079', 'CCV:00082', 'CCV:00092', 'CCV:00093', 'CCV:00096']: continue

        
    pred = hard_rules_GR(text, Evidence_type)

    if CCV in ['CCV:00008', 'CCV:00005', 'CCV:00003', 'CCV:00052', 'CCV:00065', 'CCV:00002', 'CCV:00011', 'CCV:00030']:
        if pred == None:
            continue
        else:
            print('(1)', i, text, Evidence_type, pred, CCV)
            break

    if pred != CCV:
        print('(2)', i, text, Evidence_type, pred, CCV)
        break
