==> Turkish/04. Model development.ipynb <==
# Import libraries

import warnings
warnings.filterwarnings( 'ignore' )

import os
import pickle
import joblib
import numpy    as np
from sklearn import metrics

# Data

# Load data
X = np.load('Data/data.npz', allow_pickle=True)['X']
y = np.load('Data/data.npz', allow_pickle=True)['y']

# Label encoding for targe class
from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
y = le.fit_transform(y)
pickle.dump(le, open('Model/Label_encoder.pkl', 'wb'))

## Prediction model

# import xgboost

# # Parameters
# params = {
#     'n_estimators'      : 500,
#     'learning_rate'     : 1e-1,
#     'max_depth'         : 10,
#     'reg_alpha'         : 20, 
#     'reg_lambda'        : 1.0, 
#     'gamma'             : 1, 
#     'min_child_weight'  : 2, 
#     'max_leaves'        : 2, 
#     'eval_metric'       : ['auc', 'aucpr', 'mlogloss']
# }


# # Setup model
# model = xgboost.XGBClassifier(objective           = 'multi:softmax',  
#                               n_jobs              = -1,                                   
#                               validate_parameters = True, 
#                               verbosity           = 1,
#                               tree_method         = 'hist',
#                               **params)

# # Train model
# weights = [y.shape[0] / np.where(y == i)[0].shape[0] for i in np.unique(y)]

# model.fit(X, y,
#           eval_set = [ (X, y) ],
#           sample_weight = [weights[int(x)] for x in y],
#           verbose = 50);

### Model 1

from sklearn.ensemble import RandomForestClassifier

model = RandomForestClassifier(n_estimators=300,
                               n_jobs=-1, 
                               max_depth=5, 
                               class_weight='balanced')

model.fit(X, y)

# Get predictions
pred = model.predict(X)

# Model Evaluation
Accuracy = 100*metrics.accuracy_score(y, pred)
Recall = metrics.recall_score(y, pred, average='macro')
Precision = metrics.precision_score(y, pred, average='macro')   
CM = metrics.confusion_matrix(y, pred)

print(f'Accuracy: {Accuracy:.3f}%')
print(f'Precision: {Precision:.3f}')
print(f'Recall: {Recall:.3f}')
CM

### Model 2

from sklearn.ensemble import RandomForestClassifier

model = RandomForestClassifier(n_estimators=300,
                               n_jobs=-1, 
                               max_depth=15, 
                               class_weight='balanced')

model.fit(X, y)

# Get predictions
pred = model.predict(X)

# Model Evaluation
Accuracy = 100*metrics.accuracy_score(y, pred)
Recall = metrics.recall_score(y, pred, average='macro')
Precision = metrics.precision_score(y, pred, average='macro')   
CM = metrics.confusion_matrix(y, pred)

print(f'Accuracy: {Accuracy:.3f}%')
print(f'Precision: {Precision:.3f}')
print(f'Recall: {Recall:.3f}')
CM

import joblib
joblib.dump(model, "Model/model.joblib", compress=3)
