==> Turkish/05. Inference.ipynb <==
import pickle
import joblib
import spacy
import spacyturk

# downloads the spaCyTurk model
# spacyturk.download("tr_floret_web_md")

# Load Label encoder
encoder = pickle.load(open('Model/Label_encoder.pkl', 'rb'))
# Load prediction model
model = joblib.load("Model/model.joblib")
# Load spacy model
nlp = spacy.load("tr_floret_web_md")

def inference(text):
    # Get embeddings
    embeddings = nlp(text).vector
    # Get model's prediction
    pred = model.predict( embeddings.reshape(1,-1) )
    # Return class
    return encoder.inverse_transform(pred)


inference('Gülen')
