==> Turkish/06. Evaluation.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import pickle
import joblib
import pandas as pd
import numpy as np
from sklearn import metrics
from utils.hard_rules import hard_rules

from tqdm import tqdm
tqdm.pandas()

# Import data
df = pd.read_csv('Data/_20230622-132905_training.tsv', sep='\t', header=None)

# Rename columns
df.rename(columns = {0: 'Value',
                     1: 'Evidence_type',
                     2: 'CCV'}, 
          inplace = True)

# Remove ENG instances
ENG_ccvs = [ccv for ccv in df['CCV'].unique() if 'ENG' in ccv]
df = df[ ~df['CCV'].isin(ENG_ccvs) ]

# Remove duplicates
df.drop_duplicates(inplace=True)

# Drop CCVs in which no prediction can be made
dropped_CCVs = ['CCV:00004', 'CCV:00025', 'CCV:00035', 'CCV:00036', 'CCV:00042', 'CCV:00043',
                'CCV:00047', 'CCV:00048', 'CCV:00049', 'CCV:00072', 'CCV:00084', 'CCV:00087', 'CCV:00088',
                'CCV:00079', 'CCV:00082', 'CCV:00092', 'CCV:00093', 'CCV:00096']
df = df[ ~df['CCV'].isin( dropped_CCVs )]

# Reset index
df = df.reset_index().drop('index', axis=1)

### Hard rule predictions

df['Predictions'] = df[['Value','Evidence_type']].progress_apply(lambda x: hard_rules(x.Value, x.Evidence_type), axis=1)  

# Keep hard-rule predictions
df = df[ df['Predictions'].astype('str') != "None" ]


### Model predictions

X = np.load('Data/data.npz', allow_pickle=True)['X']
y = np.load('Data/data.npz', allow_pickle=True)['y']

# Load Label encoder
encoder = pickle.load(open('Model/Label_encoder.pkl', 'rb'))
# Load prediction model
model = joblib.load("Model/model.joblib")
# Get predictions
pred = encoder.inverse_transform( model.predict(X) )


# Model Evaluation
Accuracy = 100*metrics.accuracy_score(y, pred)
Recall = metrics.recall_score(y, pred, average='macro')
Precision = metrics.precision_score(y, pred, average='macro')   
CM = metrics.confusion_matrix(y, pred)

print(f'Accuracy: {Accuracy:.3f}%')
print(f'Precision: {Precision:.3f}')
print(f'Recall: {Recall:.3f}')
CM

print( metrics.classification_report(y_true=y, y_pred=pred, target_names=np.unique(y)))

### Evaluation

Predictions = np.concatenate([df['Predictions'].values, pred])
Actuals = np.concatenate([df['CCV'].values, y])

Accuracy = 100*metrics.accuracy_score(Actuals, Predictions)
Recall = metrics.recall_score(Actuals, Predictions, average='macro')
Precision = metrics.precision_score(Actuals, Predictions, average='macro')   

print(f'Accuracy: {Accuracy:.3f}%')
print(f'Precision: {Precision:.3f}')
print(f'Recall: {Recall:.3f}')

CM = metrics.confusion_matrix(Actuals, Predictions)
CM

from utils.pretty_confusion_matrix import pp_matrix_from_data
pp_matrix_from_data(CM=CM, 
                    cmap="Oranges", 
                    figsize=(25, 25))
